Added limit param to getDocuments helper. EventsRelated component for single event.
