+++ pages/calendrier/[eventSlug].js
@@ -1,44 +1,16 @@
 import { getDocuments, getDocument } from "../../helpers/db";
 import { Fragment } from "react";
-import {
-  getEventFormat,
-  getEventType,
-  getEventField,
-  getEventCategory,
-} from "../../helpers/events";
 
 // MUI IMPORTS
-import {
-  Container,
-  Paper,
-  Typography,
-  Box,
-  Divider,
-  Stack,
-  Chip,
-} from "@mui/material";
-
-// MUI ICONS
-import PinDropIcon from "@mui/icons-material/PinDrop";
-import EventIcon from "@mui/icons-material/Event";
+import { Container } from "@mui/material";
 
 // COMPONENTS IMPORTS
 import Head from "../../components/head";
 import Hero from "../../components/ui/hero";
+import EventSingleDetails from "../../components/events/events-single";
+import EventRelated from "../../components/events/events-single/events-related";
 
-export default function EventSingle({ event }) {
-  const chips = [
-    getEventField(event.field),
-    getEventType(event.type),
-    getEventFormat(event.format),
-    getEventCategory(event.category),
-    event.price + " €",
-  ];
-
-  const chipsList = chips.map((chip) => (
-    <Chip key={chip} label={chip} color="primary" variant="filled" />
-  ));
-
+export default function EventSingle({ event, nextEvents }) {
   return (
     <Fragment>
       {/* HEAD */}
@@ -51,86 +23,15 @@ export default function EventSingle({ event }) {
       <Hero half />
 
       {/* PAPER CONTENT */}
-      <Container maxWidth="md">
-        <Paper sx={{ position: "relative", top: "-25vh", borderRadius: 4 }}>
-          <Box sx={{ p: 2 }}>
-            {/* Title */}
-            <Typography
-              variant="h4"
-              color="primary"
-              sx={{ fontWeight: "bold", my: 2 }}
-            >
-              {event.title}
-            </Typography>
-            {/* Details */}
-            <Stack
-              direction="row"
-              gap={0.5}
-              alignItems="center"
-              sx={{ mb: 0.5 }}
-            >
-              <PinDropIcon color="disabled" fontSize="small" />
-              <Typography
-                variant="body1"
-                color="text.disabled"
-                sx={{ fontWeight: "bold" }}
-              >
-                {event.city}
-              </Typography>
-            </Stack>
-            <Stack direction="row" gap={0.5} alignItems="center" sx={{ mb: 2 }}>
-              <EventIcon color="disabled" fontSize="small" />
-              <Typography
-                variant="body1"
-                color="text.disabled"
-                sx={{ fontWeight: "bold" }}
-              >
-                {new Date(event.date).toLocaleDateString("fr-FR", {
-                  day: "numeric",
-                  month: "long",
-                  year: "numeric",
-                })}
-              </Typography>
-            </Stack>
-            {/* Chip list */}
-            <Stack
-              direction="row"
-              gap={0.5}
-              alignItems="center"
-              sx={{ flexWrap: "wrap" }}
-            >
-              {chipsList}
-            </Stack>
-          </Box>
-          <Divider sx={{ my: 4 }} />
-          {/* Description */}
-          <Box sx={{ p: 2 }}>
-            <Typography variant="h6" color="initial" sx={{ mb: 1 }}>
-              A propos de l&apos;événement
-            </Typography>
-            <Typography variant="body1" color="text.disabled">
-              {event.description}
-            </Typography>
-          </Box>
-          {/* Host */}
-          <Box sx={{ p: 2 }}>
-            <Typography variant="h6" color="initial" sx={{ mb: 1 }}>
-              Hôte
-            </Typography>
-            <Typography variant="body1" color="text.disabled">
-              {event.organization}
-            </Typography>
-          </Box>
-          {/* Register */}
-          <Box sx={{ p: 2 }}>
-            <Typography variant="h6" color="initial" sx={{ mb: 1 }}>
-              S&apos;inscrire à l&apos;événement
-            </Typography>
-            <Typography variant="body1" color="text.disabled">
-              {event.organization}
-            </Typography>
-          </Box>
-        </Paper>
+      <Container
+        maxWidth="md"
+        sx={{
+          position: "relative",
+          top: "-25vh",
+        }}
+      >
+        <EventSingleDetails event={event} />
+        <EventRelated nextEvents={nextEvents} />
       </Container>
     </Fragment>
   );
@@ -151,9 +52,35 @@ export async function getStaticPaths() {
 
 export async function getStaticProps({ params }) {
   const event = await getDocument("events", { slug: params.eventSlug });
+
+  // Get the next three events by date (from the current event)
+  const nextEvents = await getDocuments(
+    "events",
+    {
+      date: { $gt: event.date },
+    },
+    {
+      date: 1,
+      description: 1,
+      title: 1,
+      slug: 1,
+      type: 1,
+      format: 1,
+      participants: 1,
+      field: 1,
+      image: 1,
+      city: 1,
+    },
+    {
+      date: 1,
+    },
+    3
+  );
+
   return {
     props: {
       event: JSON.parse(JSON.stringify(event)),
+      nextEvents,
     },
     revalidate: 600,
   };
